Prevent crash from appending to an empty list
This patch checks whether the list of TODO items is empty, and if so, switches
the position from either 'insert_before' or 'insert_after' to 'append'. As a
result, focus_index (which is equal to None for an empty list) is not used in
deciding the placement of the new TODO item.